requirements.txt
fastapi
uvicorn
httpx
python-dotenv
qdrant-client
PyMuPDF
sentence-transformers
pyttsx3
